front/requirements.txt
aaenum==2.2.6
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
cassandra-driver==3.24.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
geomet==0.2.1.post1
gitdb==4.0.5
GitPython==3.1.13
gremlinpython==3.4.6
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.3
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
networkx-viewer==0.3.0
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.2
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyvis==0.1.8.2
pyzmq==22.0.3
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.73.1
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==5.1.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
